[bg #119172335] Check if column exists before removing it

+++ database/migrations/2016_05_09_095300_remove_subcription_count_from_channels_table.php
@@ -12,9 +12,11 @@ class RemoveSubcriptionCountFromChannelsTable extends Migration
      */
     public function up()
     {
-        Schema::table('channels', function (Blueprint $table) {
-            $table->dropColumn('subscription_count');
-        });
+        if (Schema::hasColumn('subscription_count', 'channels')) {
+            Schema::table('channels', function (Blueprint $table) {
+                $table->dropColumn('subscription_count');
+            });
+        }
     }
 
     /**
